eliminate regions url path
