+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
 export default function NotFound() {
-  redirect('/regions');
+  redirect('/');
 }
